--- web/src/main.js
@@ -1,6 +1,0 @@
-'use strict';
-
-require('./css/style.css');
-require('./app.js');
-
-console.log('app started');
